refactor: create hook useSearchClient component
