+++ src/Pages/listClient.tsx
@@ -3,7 +3,8 @@ import Card from "@/components/Card/Card";
 import Spinner from "@/components/spinner/spinner";
 import useAuthentication from "@/hook/useAuthentication";
 import { SearchField } from "@/hook/useClient/types";
-import useClient, { useClientSearch } from "@/hook/useClient/useClient";
+import useClient from "@/hook/useClient/useClient";
+import { useClientSearch } from "@/hook/useClient/useSearchClient";
 import { Client } from "@/types";
 import { useState } from "react";
 import { Link } from "react-router-dom";
